feature: Delete original maddpg and rename maddpg_scaled to maddpg. fix: Fix a few small bugs.
