+++ examples/debugging/simple_spread/feedforward/decentralised/run_maddpg_scale_trainers.py
@@ -24,7 +24,7 @@ import sonnet as snt
 from absl import app, flags
 from launchpad.nodes.python.local_multi_processing import PythonProcess
 
-from mava.systems.tf import maddpg_scaled
+from mava.systems.tf import maddpg
 from mava.systems.tf.maddpg import make_default_networks
 from mava.utils import lp_utils
 from mava.utils.environments import debugging_utils
@@ -93,7 +93,7 @@ def main(_: Any) -> None:
     checkpoint_dir = f"{FLAGS.base_dir}/{FLAGS.mava_id}"
 
     # distributed program
-    program = maddpg_scaled.MADDPG(
+    program = maddpg.MADDPG(
         environment_factory=environment_factory,
         network_factory=network_factory,
         logger_config=logger_config,
