--- mava/systems/tf/maddpg/README.md
@@ -1,9 +1,0 @@
-# Multi-Agent Deep Deterministic Policy Gradient (MADDPG)
-
-This system is an extension of the Deep Deterministic Policy Gradient (DDPG) algorithm to the multi-agent setting, originally proposed by [Lowe et al. (2017)]. The system can be trained in a decentralised, centralised, networked or state-based manner.
-
-<p style="text-align:center;">
-<img src="../../../../docs/images/maddpg.png" width="40%">
-</p>
-
-[Lowe et al. (2017)]: https://arxiv.org/pdf/1706.02275
