+++ mava/systems/tf/maddpg/__init__.py
@@ -13,6 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+"""Implementations of a MADDPG agent."""
 from mava.systems.tf.maddpg.execution import (
     MADDPGFeedForwardExecutor,
     MADDPGRecurrentExecutor,
