+++ mava/systems/tf/maddpg/system.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""MADDPG system implementation."""
+"""MADDPG scaled system implementation."""
 
 import functools
 from typing import Any, Callable, Dict, List, Optional, Tuple, Type, Union
@@ -25,7 +25,7 @@ import reverb
 import sonnet as snt
 from acme import specs as acme_specs
 from acme.tf import utils as tf2_utils
-from acme.utils import counting, loggers
+from acme.utils import loggers
 
 import mava
 from mava import core
@@ -33,10 +33,9 @@ from mava import specs as mava_specs
 from mava.components.tf.architectures import DecentralisedQValueActorCritic
 from mava.environment_loop import ParallelEnvironmentLoop
 from mava.systems.tf import executors
-from mava.systems.tf import savers as tf2_savers
 from mava.systems.tf.maddpg import builder, training
 from mava.systems.tf.maddpg.execution import MADDPGFeedForwardExecutor
-from mava.utils import lp_utils
+from mava.systems.tf.variable_sources import VariableSource as MavaVariableSource
 from mava.utils.loggers import MavaLogger, logger_utils
 from mava.wrappers import DetailedPerAgentStatistics
 
@@ -58,10 +57,11 @@ class MADDPG:
         ] = training.MADDPGDecentralisedTrainer,
         executor_fn: Type[core.Executor] = MADDPGFeedForwardExecutor,
         num_executors: int = 1,
-        num_caches: int = 0,
-        environment_spec: mava_specs.MAEnvironmentSpec = None,
+        num_trainers: int = 1,
+        trainer_net_config: Dict[str, List] = {},
         shared_weights: bool = True,
         agent_net_keys: Dict[str, str] = {},
+        environment_spec: mava_specs.MAEnvironmentSpec = None,
         discount: float = 0.99,
         batch_size: int = 256,
         prefetch_size: int = 4,
@@ -81,7 +81,7 @@ class MADDPG:
         period: int = 20,
         sigma: float = 0.3,
         max_gradient_norm: float = None,
-        max_executor_steps: int = None,
+        # max_executor_steps: int = None,
         checkpoint: bool = True,
         checkpoint_subpath: str = "~/mava/",
         logger_config: Dict = {},
@@ -92,7 +92,6 @@ class MADDPG:
         connection_spec: Callable[[Dict[str, List[str]]], Dict[str, List[str]]] = None,
     ):
         """Initialise the system
-
         Args:
             environment_factory (Callable[[bool], dm_env.Environment]): function to
                 instantiate an environment.
@@ -169,7 +168,6 @@ class MADDPG:
                 the evaluation loop. Defaults to {}.
             connection_spec (Callable[[Dict[str, List[str]]], Dict[str, List[str]]],
                 optional): network topology specification for networked system
-                architectures. Defaults to None.
         """
 
         if not environment_spec:
@@ -195,14 +193,27 @@ class MADDPG:
                 for agent in agents
             }
 
+        # Setup trainer_net_config
+        self._trainer_net_config = trainer_net_config
+
+        if not trainer_net_config:
+            if num_trainers > 1:
+                raise ValueError(
+                    "Warning: For more than one trainer the "
+                    "trainer_net_config needs to be specified."
+                )
+
+            self._trainer_net_config["trainer_0"] = [
+                net_key for net_key in set(self._agent_net_keys.values())
+            ]
+
         self._architecture = architecture
         self._environment_factory = environment_factory
         self._network_factory = network_factory
         self._logger_factory = logger_factory
         self._environment_spec = environment_spec
         self._num_exectors = num_executors
-        self._num_caches = num_caches
-        self._max_executor_steps = max_executor_steps
+        self._num_trainers = num_trainers
         self._checkpoint_subpath = checkpoint_subpath
         self._checkpoint = checkpoint
         self._logger_config = logger_config
@@ -227,6 +238,8 @@ class MADDPG:
             builder.MADDPGConfig(
                 environment_spec=environment_spec,
                 agent_net_keys=self._agent_net_keys,
+                num_trainers=num_trainers,
+                num_executors=num_executors,
                 discount=discount,
                 batch_size=batch_size,
                 prefetch_size=prefetch_size,
@@ -254,7 +267,6 @@ class MADDPG:
 
     def _get_extra_specs(self) -> Any:
         """helper to establish specs for extra information
-
         Returns:
             Dict[str, Any]: dictionary containing extra specs
         """
@@ -266,26 +278,28 @@ class MADDPG:
             agent_net_keys=self._agent_net_keys,
         )
         for agent in agents:
-            agent_type = agent.split("_")[0]
+            # agent_type = agent.split("_")[0]
+            agent_net_key = self._agent_net_keys[agent]
             core_state_specs[agent] = (
                 tf2_utils.squeeze_batch_dim(
-                    networks["policies"][agent_type].initial_state(1)
+                    networks["policies"][agent_net_key].initial_state(1)
                 ),
             )
         return {"core_states": core_state_specs}
 
     def replay(self) -> Any:
-        """Replay data storage.
-
+        """Step counter
+        Args:
+            checkpoint (bool): whether to checkpoint the counter.
         Returns:
-            Any: replay data table built according the environment specification.
+            Any: step counter object.
         """
-
         return self._builder.make_replay_tables(self._environment_spec)
 
     def create_system(
         self,
     ) -> Tuple[DecentralisedQValueActorCritic, Dict[str, Dict[str, snt.Module]]]:
+        """Initialise the system variables from the network factory."""
         # Create the networks to optimize (online)
         networks = self._network_factory(  # type: ignore
             environment_spec=self._environment_spec,
@@ -310,91 +324,25 @@ class MADDPG:
         system = self._architecture(**architecture_config)
         return system, system.create_system()
 
-    def counter(self, checkpoint: bool) -> Any:
-        """Step counter
-
-        Args:
-            checkpoint (bool): whether to checkpoint the counter.
-
-        Returns:
-            Any: step counter object.
-        """
-
-        if checkpoint:
-            return tf2_savers.CheckpointingRunner(
-                counting.Counter(),
-                time_delta_minutes=15,
-                directory=self._checkpoint_subpath,
-                subdirectory="counter",
-            )
-        else:
-            return counting.Counter()
-
-    def coordinator(self, counter: counting.Counter) -> Any:
-        """Coordination helper for a distributed program
-
-        Args:
-            counter (counting.Counter): step counter object.
-
-        Returns:
-            Any: step limiter object.
-        """
-
-        return lp_utils.StepsLimiter(counter, self._max_executor_steps)
-
-    def trainer(
-        self,
-        replay: reverb.Client,
-        counter: counting.Counter,
-    ) -> mava.core.Trainer:
-        """System trainer
-
-        Args:
-            replay (reverb.Client): replay data table to pull data from.
-            counter (counting.Counter): step counter object.
-
-        Returns:
-            mava.core.Trainer: system trainer.
-        """
-
-        # create logger
-        trainer_logger_config = {}
-        if self._logger_config and "trainer" in self._logger_config:
-            trainer_logger_config = self._logger_config["trainer"]
-        trainer_logger = self._logger_factory(  # type: ignore
-            "trainer", **trainer_logger_config
-        )
-
+    def variable_server(self) -> MavaVariableSource:
+        """Create the variable server."""
         # Create the system
         _, system_networks = self.create_system()
-
-        dataset = self._builder.make_dataset_iterator(replay)
-        counter = counting.Counter(counter, "trainer")
-
-        return self._builder.make_trainer(
-            networks=system_networks,
-            dataset=dataset,
-            counter=counter,
-            logger=trainer_logger,
-            connection_spec=self._connection_spec,
-        )
+        return self._builder.make_variable_server(system_networks)
 
     def executor(
         self,
         executor_id: str,
         replay: reverb.Client,
         variable_source: acme.VariableSource,
-        counter: counting.Counter,
     ) -> mava.ParallelEnvironmentLoop:
         """System executor
-
         Args:
             executor_id (str): id to identify the executor process for logging purposes.
             replay (reverb.Client): replay data table to push data to.
             variable_source (acme.VariableSource): variable server for updating
                 network variables.
             counter (counting.Counter): step counter object.
-
         Returns:
             mava.ParallelEnvironmentLoop: environment-executor loop instance.
         """
@@ -407,6 +355,7 @@ class MADDPG:
 
         # Create the executor.
         executor = self._builder.make_executor(
+            # executor_id=executor_id,
             policy_networks=behaviour_policy_networks,
             adder=self._builder.make_adder(replay),
             variable_source=variable_source,
@@ -415,9 +364,6 @@ class MADDPG:
         # TODO (Arnu): figure out why factory function are giving type errors
         # Create the environment.
         environment = self._environment_factory(evaluation=False)  # type: ignore
-
-        # Create logger and counter; actors will not spam bigtable.
-        counter = counting.Counter(counter, "executor")
 
         # Create executor logger
         executor_logger_config = {}
@@ -431,7 +377,6 @@ class MADDPG:
         train_loop = self._train_loop_fn(
             environment,
             executor,
-            counter=counter,
             logger=exec_logger,
             **self._train_loop_fn_kwargs,
         )
@@ -443,17 +388,14 @@ class MADDPG:
     def evaluator(
         self,
         variable_source: acme.VariableSource,
-        counter: counting.Counter,
         logger: loggers.Logger = None,
     ) -> Any:
         """System evaluator (an executor process not connected to a dataset)
-
         Args:
             variable_source (acme.VariableSource): variable server for updating
                 network variables.
             counter (counting.Counter): step counter object.
             logger (loggers.Logger, optional): logger object. Defaults to None.
-
         Returns:
             Any: environment-executor evaluation loop instance for evaluating the
                 performance of a system.
@@ -467,6 +409,7 @@ class MADDPG:
 
         # Create the agent.
         executor = self._builder.make_executor(
+            # executor_id="evaluator",
             policy_networks=behaviour_policy_networks,
             variable_source=variable_source,
         )
@@ -475,7 +418,6 @@ class MADDPG:
         environment = self._environment_factory(evaluation=True)  # type: ignore
 
         # Create logger and counter.
-        counter = counting.Counter(counter, "evaluator")
         evaluator_logger_config = {}
         if self._logger_config and "evaluator" in self._logger_config:
             evaluator_logger_config = self._logger_config["evaluator"]
@@ -488,7 +430,6 @@ class MADDPG:
         eval_loop = self._eval_loop_fn(
             environment,
             executor,
-            counter=counter,
             logger=eval_logger,
             **self._eval_loop_fn_kwargs,
         )
@@ -496,55 +437,74 @@ class MADDPG:
         eval_loop = DetailedPerAgentStatistics(eval_loop)
         return eval_loop
 
+    def trainer(
+        self,
+        trainer_id: str,
+        replay: reverb.Client,
+        variable_source: MavaVariableSource,
+        # counter: counting.Counter,
+    ) -> mava.core.Trainer:
+        """System trainer
+        Args:
+            replay (reverb.Client): replay data table to pull data from.
+            counter (counting.Counter): step counter object.
+        Returns:
+            mava.core.Trainer: system trainer.
+        """
+
+        # create logger
+        trainer_logger_config = {}
+        if self._logger_config and "trainer" in self._logger_config:
+            trainer_logger_config = self._logger_config["trainer"]
+        trainer_logger = self._logger_factory(  # type: ignore
+            f"trainer_{trainer_id}", **trainer_logger_config
+        )
+
+        # Create the system
+        _, system_networks = self.create_system()
+
+        dataset = self._builder.make_dataset_iterator(replay)
+
+        return self._builder.make_trainer(
+            # trainer_id=trainer_id,
+            networks=system_networks,
+            trainer_net_config=self._trainer_net_config[f"trainer_{trainer_id}"],
+            dataset=dataset,
+            logger=trainer_logger,
+            connection_spec=self._connection_spec,
+            variable_source=variable_source,
+        )
+
     def build(self, name: str = "maddpg") -> Any:
         """Build the distributed system as a graph program.
-
         Args:
             name (str, optional): system name. Defaults to "maddpg".
-
         Returns:
             Any: graph program for distributed system training.
         """
-
         program = lp.Program(name=name)
 
         with program.group("replay"):
             replay = program.add_node(lp.ReverbNode(self.replay))
 
-        with program.group("counter"):
-            counter = program.add_node(lp.CourierNode(self.counter, self._checkpoint))
-
-        if self._max_executor_steps:
-            with program.group("coordinator"):
-                _ = program.add_node(lp.CourierNode(self.coordinator, counter))
+        with program.group("variable_server"):
+            variable_server = program.add_node(lp.CourierNode(self.variable_server))
 
         with program.group("trainer"):
-            trainer = program.add_node(lp.CourierNode(self.trainer, replay, counter))
+            # Add executors which pull round-robin from our variable sources.
+            for trainer_id in range(self._num_trainers):
+                program.add_node(
+                    lp.CourierNode(self.trainer, trainer_id, replay, variable_server)
+                )
 
         with program.group("evaluator"):
-            program.add_node(lp.CourierNode(self.evaluator, trainer, counter))
-
-        if not self._num_caches:
-            # Use the trainer as a single variable source.
-            sources = [trainer]
-        else:
-            with program.group("cacher"):
-                # Create a set of trainer caches.
-                sources = []
-                for _ in range(self._num_caches):
-                    cacher = program.add_node(
-                        lp.CacherNode(
-                            trainer, refresh_interval_ms=2000, stale_after_ms=4000
-                        )
-                    )
-                    sources.append(cacher)
+            program.add_node(lp.CourierNode(self.evaluator, variable_server))
 
         with program.group("executor"):
             # Add executors which pull round-robin from our variable sources.
             for executor_id in range(self._num_exectors):
-                source = sources[executor_id % len(sources)]
                 program.add_node(
-                    lp.CourierNode(self.executor, executor_id, replay, source, counter)
+                    lp.CourierNode(self.executor, executor_id, replay, variable_server)
                 )
 
         return program
