--- mava/systems/tf/maddpg_scaled/__init__.py
@@ -1,25 +1,0 @@
-# python3
-# Copyright 2021 InstaDeep Ltd. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Implementations of a MADDPG agent."""
-
-from mava.systems.tf.maddpg_scaled.execution import (  # MADDPGRecurrentExecutor,
-    MADDPGFeedForwardExecutor,
-)
-from mava.systems.tf.maddpg_scaled.system import MADDPG
-from mava.systems.tf.maddpg_scaled.training import (
-    MADDPGBaseTrainer,
-    MADDPGDecentralisedTrainer,
-)
