--- mava/systems/tf/maddpg_scaled/execution.py
@@ -1,184 +1,0 @@
-# python3
-# Copyright 2021 InstaDeep Ltd. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from typing import Any, Dict, List, Optional, Tuple, Union
-
-import dm_env
-import sonnet as snt
-import tensorflow as tf
-import tensorflow_probability as tfp
-from acme import types
-from acme.specs import EnvironmentSpec
-
-# Internal imports.
-from acme.tf import utils as tf2_utils
-from acme.tf import variable_utils as tf2_variable_utils
-from dm_env import specs
-
-from mava import adders
-from mava.systems.tf import executors
-
-Array = specs.Array
-BoundedArray = specs.BoundedArray
-DiscreteArray = specs.DiscreteArray
-
-tfd = tfp.distributions
-
-
-class MADDPGFeedForwardExecutor(executors.FeedForwardExecutor):
-    """A feed-forward executor for discrete actions.
-    An executor based on a feed-forward policy for each agent in the system.
-    """
-
-    def __init__(
-        self,
-        policy_networks: Dict[str, snt.Module],
-        agent_specs: Dict[str, EnvironmentSpec],
-        agent_net_keys: Dict[str, str],
-        adder: Optional[adders.ParallelAdder] = None,
-        counts: Optional[Dict[str, Any]] = None,
-        variable_client: Optional[tf2_variable_utils.VariableClient] = None,
-    ):
-
-        """Initialise the system executor
-        Args:
-            policy_networks (Dict[str, snt.Module]): policy networks for each agent in
-                the system.
-            agent_specs (Dict[str, EnvironmentSpec]): agent observation and action
-                space specifications.
-            adder (Optional[adders.ParallelAdder], optional): adder which sends data
-                to a replay buffer. Defaults to None.
-            variable_client (Optional[tf2_variable_utils.VariableClient], optional):
-                client to copy weights from the trainer. Defaults to None.
-            agent_net_keys: (dict, optional): specifies what network each agent uses.
-                Defaults to {}.
-        """
-
-        # Store these for later use.
-        self._agent_specs = agent_specs
-        self._counts = counts
-        super().__init__(
-            policy_networks=policy_networks,
-            agent_net_keys=agent_net_keys,
-            adder=adder,
-            variable_client=variable_client,
-        )
-
-    @tf.function
-    def _policy(
-        self, agent: str, observation: types.NestedTensor
-    ) -> types.NestedTensor:
-        """Agent specific policy function
-
-        Args:
-            agent (str): agent id
-            observation (types.NestedTensor): observation tensor received from the
-                environment.
-
-        Raises:
-            NotImplementedError: unknown action space
-
-        Returns:
-            types.NestedTensor: agent action
-        """
-
-        # Add a dummy batch dimension and as a side effect convert numpy to TF.
-        batched_observation = tf2_utils.add_batch_dim(observation)
-
-        # index network either on agent type or on agent id
-        agent_key = self._agent_net_keys[agent]
-
-        # Compute the policy, conditioned on the observation.
-        policy = self._policy_networks[agent_key](batched_observation)
-
-        # TODO (dries): Make this support hybrid action spaces.
-        if type(self._agent_specs[agent].actions) == BoundedArray:
-            # Continuous action
-            action = policy
-        elif type(self._agent_specs[agent].actions) == DiscreteArray:
-            action = tf.math.argmax(policy, axis=1)
-        else:
-            raise NotImplementedError
-
-        return action, policy
-
-    def select_action(
-        self, agent: str, observation: types.NestedArray
-    ) -> Tuple[types.NestedArray, types.NestedArray]:
-        """select an action for a single agent in the system
-
-        Args:
-            agent (str): agent id.
-            observation (types.NestedArray): observation tensor received from the
-                environment.
-
-        Returns:
-            Tuple[types.NestedArray, types.NestedArray]: agent action and policy.
-        """
-        # Step the recurrent policy/value network forward
-        # given the current observation and state.
-        action, policy = self._policy(agent, observation.observation)
-
-        # Return a numpy array with squeezed out batch dimension.
-        action = tf2_utils.to_numpy_squeeze(action)
-        policy = tf2_utils.to_numpy_squeeze(policy)
-        return action, policy
-
-    def select_actions(
-        self, observations: Dict[str, types.NestedArray]
-    ) -> Tuple[Dict[str, types.NestedArray], Dict[str, types.NestedArray]]:
-        """select the actions for all agents in the system
-
-        Args:
-            observations (Dict[str, types.NestedArray]): agent observations from the
-                environment.
-
-        Returns:
-            Tuple[Dict[str, types.NestedArray], Dict[str, types.NestedArray]]:
-                actions and policies for all agents in the system.
-        """
-
-        actions = {}
-        policies = {}
-        for agent, observation in observations.items():
-            actions[agent], policies[agent] = self.select_action(agent, observation)
-        return actions, policies
-
-    def observe(
-        self,
-        actions: Union[
-            Dict[str, types.NestedArray], List[Dict[str, types.NestedArray]]
-        ],
-        next_timestep: dm_env.TimeStep,
-        next_extras: Dict[str, types.NestedArray] = {},
-    ) -> None:
-        """record observed timestep from the environment
-        Args:
-            actions (Union[ Dict[str, types.NestedArray], List[Dict[str,
-                types.NestedArray]] ]): system agents' actions.
-            next_timestep (dm_env.TimeStep): data emitted by an environment during
-                interaction.
-            next_extras (Dict[str, types.NestedArray], optional): possible extra
-                information to record during the transition. Defaults to {}.
-        """
-        if self._adder:
-            _, policy = actions
-            # TODO (dries): Sort out this mypy issue.
-            self._adder.add(policy, next_timestep, next_extras)  # type: ignore
-
-    def update(self, wait: bool = False) -> None:
-        if self._variable_client:
-            # TODO (dries): Maybe changes this to a async get?
-            self._variable_client.get_and_wait()
