--- mava/systems/tf/maddpg_scaled/training.py
@@ -1,562 +1,0 @@
-# python3
-# Copyright 2021 InstaDeep Ltd. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-"""MADDPG trainer implementation."""
-
-import copy
-import time
-from typing import Any, Dict, List, Optional, Sequence, Tuple, Union
-
-import numpy as np
-import reverb
-import sonnet as snt
-import tensorflow as tf
-import tree
-import trfl
-from acme.tf import losses
-from acme.tf import utils as tf2_utils
-from acme.utils import loggers
-
-from mava.systems.tf.maddpg.training import MADDPGBaseTrainer as feedforward_trainer
-from mava.systems.tf.variable_utils import VariableClient
-from mava.utils import training_utils as train_utils
-
-train_utils.set_growing_gpu_memory()
-
-
-class MADDPGBaseTrainer(feedforward_trainer):
-    """MADDPG trainer.
-    This is the trainer component of a MADDPG system. IE it takes a dataset as input
-    and implements update functionality to learn from this dataset.
-    """
-
-    def __init__(
-        self,
-        agents: List[str],
-        agent_types: List[str],
-        trainer_net_config: List[str],
-        policy_networks: Dict[str, snt.Module],
-        critic_networks: Dict[str, snt.Module],
-        target_policy_networks: Dict[str, snt.Module],
-        target_critic_networks: Dict[str, snt.Module],
-        policy_optimizer: Union[snt.Optimizer, Dict[str, snt.Optimizer]],
-        critic_optimizer: Union[snt.Optimizer, Dict[str, snt.Optimizer]],
-        discount: float,
-        target_averaging: bool,
-        target_update_period: int,
-        target_update_rate: float,
-        dataset: tf.data.Dataset,
-        observation_networks: Dict[str, snt.Module],
-        target_observation_networks: Dict[str, snt.Module],
-        variable_client: VariableClient,
-        counts: Dict[str, Any],
-        num_steps: int,
-        agent_net_keys: Dict[str, str],
-        max_gradient_norm: float = None,
-        logger: loggers.Logger = None,
-    ):
-        """Initialise MADDPG trainer
-        Args:
-            agents (List[str]): agent ids, e.g. "agent_0".
-            agent_types (List[str]): agent types, e.g. "speaker" or "listener".
-            policy_networks (Dict[str, snt.Module]): policy networks for each agent in
-                the system.
-            critic_networks (Dict[str, snt.Module]): critic network(s), shared or for
-                each agent in the system.
-            target_policy_networks (Dict[str, snt.Module]): target policy networks.
-            target_critic_networks (Dict[str, snt.Module]): target critic networks.
-            policy_optimizer (Union[snt.Optimizer, Dict[str, snt.Optimizer]]):
-                optimizer(s) for updating policy networks.
-            critic_optimizer (Union[snt.Optimizer, Dict[str, snt.Optimizer]]):
-                optimizer for updating critic networks.
-            discount (float): discount factor for TD updates.
-            target_averaging (bool): whether to use polyak averaging for target network
-                updates.
-            target_update_period (int): number of steps before target networks are
-                updated.
-            target_update_rate (float): update rate when using averaging.
-            dataset (tf.data.Dataset): training dataset.
-            observation_networks (Dict[str, snt.Module]): network for feature
-                extraction from raw observation.
-            target_observation_networks (Dict[str, snt.Module]): target observation
-                network.
-            agent_net_keys: (dict, optional): specifies what network each agent uses.
-                Defaults to {}.
-            max_gradient_norm (float, optional): maximum allowed norm for gradients
-                before clipping is applied. Defaults to None.
-            counter (counting.Counter, optional): step counter object. Defaults to None.
-            logger (loggers.Logger, optional): logger object for logging trainer
-                statistics. Defaults to None.
-            checkpoint (bool, optional): whether to checkpoint networks. Defaults to
-                True.
-            checkpoint_subpath (str, optional): subdirectory for storing checkpoints.
-                Defaults to "~/mava/".
-        """
-
-        self._agents = agents
-        self._agent_types = agent_types
-        self._agent_net_keys = agent_net_keys
-        self._variable_client = variable_client
-
-        # Setup counts
-        self._counts = counts
-
-        # Store online and target networks.
-        self._policy_networks = policy_networks
-        self._critic_networks = critic_networks
-        self._target_policy_networks = target_policy_networks
-        self._target_critic_networks = target_critic_networks
-
-        # Ensure obs and target networks are sonnet modules
-        self._observation_networks = {
-            k: tf2_utils.to_sonnet_module(v) for k, v in observation_networks.items()
-        }
-        self._target_observation_networks = {
-            k: tf2_utils.to_sonnet_module(v)
-            for k, v in target_observation_networks.items()
-        }
-
-        # General learner book-keeping and loggers.
-        self._logger = logger or loggers.make_default_logger("trainer")
-
-        # Other learner parameters.
-        self._discount = discount
-
-        # Set up gradient clipping.
-        if max_gradient_norm is not None:
-            self._max_gradient_norm = tf.convert_to_tensor(max_gradient_norm)
-        else:  # A very large number. Infinity results in NaNs.
-            self._max_gradient_norm = tf.convert_to_tensor(1e10)
-
-        # Necessary to track when to update target networks.
-        self._num_steps = num_steps
-        self._target_averaging = target_averaging
-        self._target_update_period = target_update_period
-        self._target_update_rate = target_update_rate
-
-        # Create an iterator to go through the dataset.
-        self._iterator = iter(dataset)  # pytype: disable=wrong-arg-types
-
-        # Dictionary with unique network keys.
-        self.unique_net_keys = set(self._agent_net_keys.values())
-
-        # Get the agents which shoud be updated and ran
-        self._trainer_agent_list = []
-        for agent in self._agents:
-            agent_key = self._agent_net_keys[agent]
-            if agent_key in trainer_net_config:
-                self._trainer_agent_list.append(agent)
-
-        # Create optimizers for different agent types.
-        if not isinstance(policy_optimizer, dict):
-            self._policy_optimizers: Dict[str, snt.Optimizer] = {}
-            for agent in self.unique_net_keys:
-                self._policy_optimizers[agent] = copy.deepcopy(policy_optimizer)
-        else:
-            self._policy_optimizers = policy_optimizer
-
-        self._critic_optimizers: Dict[str, snt.Optimizer] = {}
-        for agent in self.unique_net_keys:
-            self._critic_optimizers[agent] = copy.deepcopy(critic_optimizer)
-
-        # Expose the variables.
-        policy_networks_to_expose = {}
-        self._system_network_variables: Dict[str, Dict[str, snt.Module]] = {
-            "critics": {},
-            "policies": {},
-        }
-        for agent_key in self.unique_net_keys:
-            policy_network_to_expose = snt.Sequential(
-                [
-                    self._target_observation_networks[agent_key],
-                    self._target_policy_networks[agent_key],
-                ]
-            )
-            policy_networks_to_expose[agent_key] = policy_network_to_expose
-            self._system_network_variables["critics"][
-                agent_key
-            ] = target_critic_networks[agent_key].variables
-            self._system_network_variables["policies"][
-                agent_key
-            ] = policy_network_to_expose.variables
-
-        # Do not record timestamps until after the first learning step is done.
-        # This is to avoid including the time it takes for actors to come online and
-        # fill the replay buffer.
-        self._timestamp: Optional[float] = None
-
-    def _update_target_networks(self) -> None:
-        for key in self.unique_net_keys:
-            # Update target network.
-            online_variables = (
-                *self._observation_networks[key].variables,
-                *self._critic_networks[key].variables,
-                *self._policy_networks[key].variables,
-            )
-            target_variables = (
-                *self._target_observation_networks[key].variables,
-                *self._target_critic_networks[key].variables,
-                *self._target_policy_networks[key].variables,
-            )
-
-            if self._target_averaging:
-                assert 0.0 < self._target_update_rate < 1.0
-                tau = self._target_update_rate
-                for src, dest in zip(online_variables, target_variables):
-                    dest.assign(dest * (1.0 - tau) + src * tau)
-            else:
-                # Make online -> target network update ops.
-                if tf.math.mod(self._num_steps, self._target_update_period) == 0:
-                    for src, dest in zip(online_variables, target_variables):
-                        dest.assign(src)
-            self._num_steps.assign_add(1)
-
-    def get_variables(self, names: Sequence[str]) -> Dict[str, Dict[str, np.ndarray]]:
-        variables: Dict[str, Dict[str, np.ndarray]] = {}
-        for network_type in names:
-            variables[network_type] = {
-                agent: tf2_utils.to_numpy(
-                    self._system_network_variables[network_type][agent]
-                )
-                for agent in self.unique_net_keys
-            }
-        return variables
-
-    def _transform_observations(
-        self, obs: Dict[str, np.ndarray], next_obs: Dict[str, np.ndarray]
-    ) -> Tuple[Dict[str, np.ndarray], Dict[str, np.ndarray]]:
-        o_tm1 = {}
-        o_t = {}
-        for agent in self._agents:
-            agent_key = self._agent_net_keys[agent]
-            o_tm1[agent] = self._observation_networks[agent_key](obs[agent].observation)
-            o_t[agent] = self._target_observation_networks[agent_key](
-                next_obs[agent].observation
-            )
-            # This stop_gradient prevents gradients to propagate into the target
-            # observation network. In addition, since the online policy network is
-            # evaluated at o_t, this also means the policy loss does not influence
-            # the observation network training.
-            o_t[agent] = tree.map_structure(tf.stop_gradient, o_t[agent])
-        return o_tm1, o_t
-
-    def _get_critic_feed(
-        self,
-        o_tm1_trans: Dict[str, np.ndarray],
-        o_t_trans: Dict[str, np.ndarray],
-        a_tm1: Dict[str, np.ndarray],
-        a_t: Dict[str, np.ndarray],
-        e_tm1: Dict[str, np.ndarray],
-        e_t: Dict[str, np.array],
-        agent: str,
-    ) -> Tuple[tf.Tensor, tf.Tensor, tf.Tensor, tf.Tensor]:
-        """get data to feed to the agent critic network(s)
-
-        Args:
-            o_tm1_trans (Dict[str, np.ndarray]): transformed (e.g. using observation
-                network) observation at timestep t-1
-            o_t_trans (Dict[str, np.ndarray]): transformed observation at timestep t
-            a_tm1 (Dict[str, np.ndarray]): action at timestep t-1
-            a_t (Dict[str, np.ndarray]): action at timestep t
-            e_tm1 (Dict[str, np.ndarray]): extras at timestep t-1
-            e_t (Dict[str, np.array]): extras at timestep t
-            agent (str): agent id
-
-        Returns:
-            Tuple[tf.Tensor, tf.Tensor, tf.Tensor, tf.Tensor]: agent critic network
-                feeds
-        """
-
-        # Decentralised critic
-        o_tm1_feed = o_tm1_trans[agent]
-        o_t_feed = o_t_trans[agent]
-        a_tm1_feed = a_tm1[agent]
-        a_t_feed = a_t[agent]
-        return o_tm1_feed, o_t_feed, a_tm1_feed, a_t_feed
-
-    def _get_dpg_feed(
-        self,
-        a_t: Dict[str, np.ndarray],
-        dpg_a_t: np.ndarray,
-        agent: str,
-    ) -> tf.Tensor:
-        """get data to feed to the agent networks
-
-        Args:
-            a_t (Dict[str, np.ndarray]): action at timestep t
-            dpg_a_t (np.ndarray): predicted action at timestep t
-            agent (str): agent id
-
-        Returns:
-            tf.Tensor: agent policy network feed
-        """
-        # Decentralised DPG
-        dpg_a_t_feed = dpg_a_t
-        return dpg_a_t_feed
-
-    def _target_policy_actions(self, next_obs: Dict[str, np.ndarray]) -> Any:
-        """select actions using target policy networks
-
-        Args:
-            next_obs (Dict[str, np.ndarray]): next agent observations.
-
-        Returns:
-            Any: agent target actions
-        """
-        actions = {}
-        for agent in self._agents:
-            agent_key = self._agent_net_keys[agent]
-            next_observation = next_obs[agent]
-            actions[agent] = self._target_policy_networks[agent_key](next_observation)
-        return actions
-
-    @tf.function
-    def _step(
-        self,
-    ) -> Dict[str, Dict[str, Any]]:
-        """Trainer forward and backward passes.
-
-        Returns:
-            Dict[str, Dict[str, Any]]: losses
-        """
-
-        # Update the target networks
-        self._update_target_networks()
-
-        # Draw a batch of data from replay.
-        sample: reverb.ReplaySample = next(self._iterator)
-
-        self._forward(sample)
-
-        self._backward()
-
-        # Log losses per agent
-        return train_utils.map_losses_per_agent_ac(
-            self.critic_losses, self.policy_losses
-        )
-
-    # Forward pass that calculates loss.
-    def _forward(self, inputs: Any) -> None:
-        """Trainer forward pass
-        Args:
-            inputs (Any): input data from the data table (transitions)
-        """
-
-        # Unpack input data as follows:
-        # o_tm1 = dictionary of observations one for each agent
-        # a_tm1 = dictionary of actions taken from obs in o_tm1
-        # e_tm1 [Optional] = extra data for timestep t-1
-        # that the agents persist in replay.
-        # r_t = dictionary of rewards or rewards sequences
-        #   (if using N step transitions) ensuing from actions a_tm1
-        # d_t = environment discount ensuing from actions a_tm1.
-        #   This discount is applied to future rewards after r_t.
-        # o_t = dictionary of next observations or next observation sequences
-        # e_t [Optional] = extra data for timestep t that the agents persist in replay.
-        o_tm1, a_tm1, e_tm1, r_t, d_t, o_t, e_t = inputs.data
-
-        self.policy_losses = {}
-        self.critic_losses = {}
-
-        # Do forward passes through the networks and calculate the losses
-        with tf.GradientTape(persistent=True) as tape:
-
-            o_tm1_trans, o_t_trans = self._transform_observations(o_tm1, o_t)
-            a_t = self._target_policy_actions(o_t_trans)
-            for agent in self._trainer_agent_list:
-                agent_key = self._agent_net_keys[agent]
-
-                # Get critic feed
-                o_tm1_feed, o_t_feed, a_tm1_feed, a_t_feed = self._get_critic_feed(
-                    o_tm1_trans=o_tm1_trans,
-                    o_t_trans=o_t_trans,
-                    a_tm1=a_tm1,
-                    a_t=a_t,
-                    e_tm1=e_tm1,
-                    e_t=e_t,
-                    agent=agent,
-                )
-
-                # Critic learning.
-                q_tm1 = self._critic_networks[agent_key](o_tm1_feed, a_tm1_feed)
-                q_t = self._target_critic_networks[agent_key](o_t_feed, a_t_feed)
-
-                # Squeeze into the shape expected by the td_learning implementation.
-                q_tm1 = tf.squeeze(q_tm1, axis=-1)  # [B]
-                q_t = tf.squeeze(q_t, axis=-1)  # [B]
-
-                # Cast the additional discount to match the environment discount dtype.
-                discount = tf.cast(self._discount, dtype=d_t[agent].dtype)
-
-                # Critic loss.
-                critic_loss = trfl.td_learning(
-                    q_tm1, r_t[agent], discount * d_t[agent], q_t
-                ).loss
-                self.critic_losses[agent] = tf.reduce_mean(critic_loss, axis=0)
-
-                # Actor learning.
-                o_t_agent_feed = o_t_trans[agent]
-                dpg_a_t = self._policy_networks[agent_key](o_t_agent_feed)
-
-                # Get dpg actions
-                dpg_a_t_feed = self._get_dpg_feed(a_t, dpg_a_t, agent)
-
-                # Get dpg Q values.
-                dpg_q_t = self._critic_networks[agent_key](o_t_feed, dpg_a_t_feed)
-
-                # Actor loss. If clipping is true use dqda clipping and clip the norm.
-                dqda_clipping = 1.0 if self._max_gradient_norm is not None else None
-                clip_norm = self._max_gradient_norm is not None
-
-                policy_loss = losses.dpg(
-                    dpg_q_t,
-                    dpg_a_t,
-                    tape=tape,
-                    dqda_clipping=dqda_clipping,
-                    clip_norm=clip_norm,
-                )
-
-                self.policy_losses[agent] = tf.reduce_mean(policy_loss, axis=0)
-        self.tape = tape
-
-    # Backward pass that calculates gradients and updates network.
-    def _backward(self) -> None:
-        """Trainer backward pass updating network parameters"""
-
-        # Calculate the gradients and update the networks
-        policy_losses = self.policy_losses
-        critic_losses = self.critic_losses
-        tape = self.tape
-        for agent in self._trainer_agent_list:
-            agent_key = self._agent_net_keys[agent]
-
-            # Get trainable variables.
-            policy_variables = (
-                self._observation_networks[agent_key].trainable_variables
-                + self._policy_networks[agent_key].trainable_variables
-            )
-            critic_variables = (
-                # In this agent, the critic loss trains the observation network.
-                self._observation_networks[agent_key].trainable_variables
-                + self._critic_networks[agent_key].trainable_variables
-            )
-
-            # Compute gradients.
-            # Note: Warning "WARNING:tensorflow:Calling GradientTape.gradient
-            #  on a persistent tape inside its context is significantly less efficient
-            #  than calling it outside the context." caused by losses.dpg, which calls
-            #  tape.gradient.
-            policy_gradients = tape.gradient(policy_losses[agent], policy_variables)
-            critic_gradients = tape.gradient(critic_losses[agent], critic_variables)
-
-            # Maybe clip gradients.
-            policy_gradients = tf.clip_by_global_norm(
-                policy_gradients, self._max_gradient_norm
-            )[0]
-            critic_gradients = tf.clip_by_global_norm(
-                critic_gradients, self._max_gradient_norm
-            )[0]
-
-            # Apply gradients.
-            self._policy_optimizers[agent_key].apply(policy_gradients, policy_variables)
-            self._critic_optimizers[agent_key].apply(critic_gradients, critic_variables)
-        train_utils.safe_del(self, "tape")
-
-    def step(self) -> None:
-        """trainer step to update the parameters of the agents in the system"""
-
-        # Run the learning step.
-        fetches = self._step()
-
-        # Compute elapsed time.
-        timestamp = time.time()
-        elapsed_time = timestamp - self._timestamp if self._timestamp else 0
-        self._timestamp = timestamp
-
-        # Update our counts and record it.
-        # TODO (dries): Can this be simplified? Only one set and one get?
-        self._variable_client.add_and_wait(
-            ["trainer_steps", "trainer_walltime"],
-            {"trainer_steps": 1, "trainer_walltime": elapsed_time},
-        )
-
-        # Update the variable source and the trainer
-        # TODO (dries): Can this be simplified? Do an async set and get?
-        self._variable_client.set_and_wait()
-        self._variable_client.get_and_wait()
-
-        raise NotImplementedError("A trainer statistics wrapper should overwrite this.")
-
-        # Checkpoint and attempt to write the logs.
-        if self._checkpoint:
-            train_utils.checkpoint_networks(self._system_checkpointer)
-
-        if self._logger:
-            self._logger.write(fetches)
-
-
-class MADDPGDecentralisedTrainer(MADDPGBaseTrainer):
-    """MADDPG trainer for a decentralised architecture."""
-
-    def __init__(
-        self,
-        agents: List[str],
-        agent_types: List[str],
-        trainer_net_config: List[str],
-        policy_networks: Dict[str, snt.Module],
-        critic_networks: Dict[str, snt.Module],
-        target_policy_networks: Dict[str, snt.Module],
-        target_critic_networks: Dict[str, snt.Module],
-        policy_optimizer: Union[snt.Optimizer, Dict[str, snt.Optimizer]],
-        critic_optimizer: snt.Optimizer,
-        discount: float,
-        target_averaging: bool,
-        target_update_period: int,
-        target_update_rate: float,
-        dataset: tf.data.Dataset,
-        observation_networks: Dict[str, snt.Module],
-        target_observation_networks: Dict[str, snt.Module],
-        variable_client: VariableClient,
-        counts: Dict[str, Any],
-        num_steps: int,
-        agent_net_keys: Dict[str, str],
-        max_gradient_norm: float = None,
-        logger: loggers.Logger = None,
-    ):
-        super().__init__(
-            agents=agents,
-            agent_types=agent_types,
-            trainer_net_config=trainer_net_config,
-            policy_networks=policy_networks,
-            critic_networks=critic_networks,
-            target_policy_networks=target_policy_networks,
-            target_critic_networks=target_critic_networks,
-            discount=discount,
-            target_averaging=target_averaging,
-            target_update_period=target_update_period,
-            target_update_rate=target_update_rate,
-            dataset=dataset,
-            observation_networks=observation_networks,
-            target_observation_networks=target_observation_networks,
-            agent_net_keys=agent_net_keys,
-            policy_optimizer=policy_optimizer,
-            critic_optimizer=critic_optimizer,
-            max_gradient_norm=max_gradient_norm,
-            logger=logger,
-            variable_client=variable_client,
-            counts=counts,
-            num_steps=num_steps,
-        )
